+++ tests/test_compiler.py
@@ -1,6 +1,7 @@
 import pytest
 import numpy as np
 from numpy.testing import assert_array_equal
+from scipy import integrate
 from qutip_qip.compiler.gatecompiler import _default_window_t_max
 
 from qutip_qip.device import (
@@ -39,11 +40,57 @@ def test_compiling_with_scheduler():
     assert(abs(time_scheduled2 * 2 - time_not_scheduled) < 1.0e-10)
 
 
-def gauss_dist(t, sigma, amplitude, duration):
-    return amplitude/np.sqrt(2*np.pi) /sigma*np.exp(-0.5*((t-duration/2)/sigma)**2)
+def test_compiling_gates_different_sampling_number():
+    """
+    Define compiler without a physical model.
+    Test compiling gates to pulses with different sampling number.
+    """
+
+    class MockCompiler(GateCompiler):
+        def __init__(self, num_qubits, params=None):
+            super().__init__(num_qubits, params=params)
+            self.gate_compiler["U1"] = self.single_qubit_gate_compiler
+            self.gate_compiler["U2"] = self.two_qubit_gate_compiler
+            self.args.update({"params": params})
+
+        def single_qubit_gate_compiler(self, gate, args):
+            pulse_info = [("x", np.array([1.0] * 3))]
+            return [
+                Instruction(
+                    gate, tlist=np.linspace(0, 2, 3), pulse_info=pulse_info
+                )
+            ]
+
+        def two_qubit_gate_compiler(self, gate, args):
+            pulse_info = [("xx", np.array([2.0] * 5))]
+            return [
+                Instruction(
+                    gate, tlist=np.linspace(0, 4, 5), pulse_info=pulse_info
+                )
+            ]
+
+    num_qubits = 2
+    circuit = QubitCircuit(num_qubits)
+    circuit.add_gate("U1", targets=0, arg_value=1.0)
+    circuit.add_gate("U2", targets=[0, 1], arg_value=1.0)
+    circuit.add_gate("U1", targets=0, arg_value=1.0)
+
+    compiler = MockCompiler(num_qubits=2)
+    compiled_tlists, compiled_coeffs = compiler.compile(circuit)
+
+    # Filter out the nonzero part of the pulse
+    # and check if they are correct.
+    np.testing.assert_array_equal(
+        compiled_tlists["x"][np.nonzero(compiled_coeffs["x"])[0]],
+        np.array([1, 2, 7, 8]),
+    )
+    np.testing.assert_array_equal(
+        compiled_tlists["xx"][np.nonzero(compiled_coeffs["xx"])[0]],
+        np.array([3, 4, 5, 6]),
+    )
 
 
-from qutip_qip.compiler import GateCompiler
+# Test the compiler with a physical model.
 class MyCompiler(GateCompiler):  # compiler class
     def __init__(self, num_qubits, params):
         super(MyCompiler, self).__init__(num_qubits, params=params)
@@ -155,22 +202,11 @@ def test_compiler_result_format():
 
 
 @pytest.mark.parametrize(
-    "shape", list(_default_window_t_max.keys()) + ["rectangular"])
-def test_pulse_shape_scipy(shape):
-    """Test different pulse shape functions imported from scipy"""
-    num_qubits = 1
-    circuit = QubitCircuit(num_qubits)
-    circuit.add_gate("X", 0)
-    processor = LinearSpinChain(num_qubits)
-    compiler = SpinChainCompiler(num_qubits, processor.params)
-    compiler.args.update({"shape": shape, "num_samples": 100})
-    processor.load_circuit(circuit, compiler=compiler)
-    if shape == "rectangular":
-        processor.pulse_mode = "discrete"
-    else:
-        processor.pulse_mode = "continuous"
-    init_state = basis(2, 0)
-    num_result = processor.run_state(init_state).states[-1]
-    ideal_result = circuit.run(init_state)
-    ifid = 1 - fidelity(num_result, ideal_result)
-    assert(pytest.approx(ifid, abs=0.01) == 0)
+    "shape", list(_default_window_t_max.keys()))
+def test_pulse_shape(shape):
+    """Test different pulse shape functions"""
+    coeff, tlist = GateCompiler.generate_pulse_shape(
+        shape, 1001, maximum=1.0, area=1.0)
+    assert pytest.approx(coeff[500], 1.e-2) == 1  # max
+    result = integrate.trapz(coeff, tlist)
+    assert pytest.approx(result, rel=1.e-2) == 1  # area