+++ doc/source/changelog.rst
@@ -2,6 +2,30 @@
 Changelog
 *********
 
+
+Version 0.2.3 (December 12, 2022)
++++++++++++++++++++++++++++++++++
+
+Improvements
+------------
+- Add DRAG correction for the ``SCQubits`` simulator. Also improve the interpolation for the idling gates for continuous pulse shape. (`#156 <https://github.com/qutip/qutip-qip/pull/156>`_)
+
+Bug Fixes
+---------
+- Fix usage of ``rand_unitary`` and ``rand_ket``. (`#163 <https://github.com/qutip/qutip-qip/pull/163>`_)
+- Fix the final time calculation in ``GateCompiler.compile``. (`#162 <https://github.com/qutip/qutip-qip/pull/162>`_)
+- Fix bug for circuit with selected measurement result. (`#167 <https://github.com/qutip/qutip-qip/pull/167>`_)
+- Fix bugs in the analytical pulse shapes. (`#169 <https://github.com/qutip/qutip-qip/pull/169>`_)
+- Fix the swap gate plotting. (`#170 <https://github.com/qutip/qutip-qip/pull/170>`_) and (`#177 <https://github.com/qutip/qutip-qip/pull/177>`_)
+- Filter the measurement probability (`#182 <https://github.com/qutip/qutip-qip/pull/182>`_)
+- Add a default max_step based on the circuit time (`#183 <https://github.com/qutip/qutip-qip/pull/183>`_)
+
+Miscellaneous
+-------------
+- Fix parsing of package versions in packaging 22. (`#181 <https://github.com/qutip/qutip-qip/pull/181>`_)
+
+
+
 Version 0.2.2 (June 20, 2022)
 +++++++++++++++++++++++++++++